added faux chain
Because Aries does not support a chain of events like Ethereum, I have a fake chain that is just a text file.
The idea is that after a User sends PII to an FSP and the FSP verifies it and the FSP sends back a credential,
the FSP will publish that it has sent the credential on the chain. When the User presents that credential
to another FSP, that FSP will check the chain for validity of that credential, thus preempting verification time and costs.